Layout and styling updates + reload bug fix
