+++ code/src/pages/MoviesDetails.css
@@ -75,11 +75,3 @@
 .moviedetails-content a:hover {
   color: rgba(255, 255, 255, 0.5);
 }
-
-@media (max-width: 1023px) {
-
-}
-
-@media (max-width: 500px) {
-
-}
