+++ code/src/pages/MoviesList.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { MovieThumb } from './MovieThumb';
+import { MovieThumb } from '../components/MovieThumb';
 
 export const MoviesList = ({ category }) => {
   const MOVIES_URL = `https://api.themoviedb.org/3/movie/${category}?api_key=1eadd8601c20e4a3ebe5e0eca940633f&language=en-US&page=1`;
